Prince William never wears his wedding ring: lets disscuion the Reason
In 2010, the engagement of Prince William and Kate was announced And they gifted Kate a ring at the engagement. Princess Catherine has never been seen without jewelry, but recently she was seen without jewelry and a ring. When asked about it, he said it was his personal preference. This ring was made of gold which was worn by him in engagement. William and Kate chose Wartski to make their wedding ring.